tests/requirements.txt
pytest>=7.4.0
pytest-testinfra>=10.0.0
